FDG-10264 Add helper and legend toggle
