+++ src/components/data-preview/data-preview-chart/chart-legend-panel/chart-legend-panel.module.scss
@@ -14,7 +14,7 @@
 
 .buttonContainer {
   padding: 1rem;
-  max-height: 470px;
+  max-height: 372px;
   overflow-y: auto;
 }
