+++ src/components/data-preview/data-preview-chart/data-preview-chart.jsx
@@ -3,6 +3,7 @@ import drawChart from '../../charts/chart-primary';
 import ChartLegend from './chart-legend/chart-legend';
 import { thinDataAsNeededForChart } from '../../dataset-data/dataset-data-helper/dataset-data-helper';
 import {
+  billionsLabel,
   chartArea,
   chartLegendWrapper,
   chartPane,
@@ -11,72 +12,33 @@ import {
   labelContainer,
   legend as legendClass,
   legendActive,
+  legendToggle,
   viz as vizClass,
   yAxisLabel,
 } from './data-preview-chart.module.scss';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faInfoCircle } from '@fortawesome/free-solid-svg-icons';
 import ChartLegendPanel from './chart-legend-panel/chart-legend-panel';
+import { callbacks, dataTableChartNotesText, determineFormat, setFieldsToChart } from './chart-helper';
+import ChartCitation from '../../dataset-data/dataset-chart/chart-citation/chart-citation';
 
 export let chartHooks;
-export const callbacks = {
-  onHover: (on, item, hasUpdates, chartFields) => {
-    const isActiveField = chartFields.some(field => field.field === item.field && field.active);
-    if (isActiveField && chartHooks.onHover && hasUpdates) {
-      chartHooks.onHover(on, item.field);
-    }
-  },
-  onLabelChange: (update, chartFields, setChartFields) => {
-    if (!(chartFields && setChartFields && chartHooks.onFieldUpdates)) {
-      return;
-    }
 
-    const selectedFields = update.map(r => r.field);
+/*
+TODO
+- data props
+- legend button
+- style footer buttons
+- legend panel height
+ */
 
-    chartFields.forEach(r => {
-      r.active = selectedFields.indexOf(r.field) !== -1;
-    });
-
-    chartHooks.onFieldUpdates(selectedFields);
-    setChartFields(chartFields.slice());
-  },
-};
-
-export const setFieldsToChart = (fields, pivot) => {
-  const whiteList = ['currency', 'number', 'percentage'];
-  const blackList = ['src_line_nbr', 'table_nbr', 'total_incoming_transfers_cnt', 'from_legacy_system_cnt', 'from_commercial_book_entry_cnt'];
-  const filteredChartFields = Object.keys(fields).filter(
-    f =>
-      (whiteList.indexOf(fields[f].toLowerCase()) !== -1 || whiteList.indexOf(fields[f].substring(0, 8).toLowerCase()) !== -1) &&
-      blackList.indexOf(f.toLowerCase()) === -1
-  );
-  // pivot has synthetic columns, so order them alphabetically for display in the legend
-  if (pivot && pivot.pivotView && pivot.pivotView.dimensionField) {
-    return filteredChartFields.sort((a, b) => a.localeCompare(b));
-  } else {
-    return filteredChartFields;
-  }
-};
-
-export const determineFormat = (fields, dataTypes) => {
-  const isPercentage = fields.every(f => dataTypes[f].toLowerCase() === 'percentage');
-  const isCurrency = fields.every(f => dataTypes[f].toLowerCase() === 'currency' || dataTypes[f].substring(0, 8).toLowerCase() === 'currency');
-  return isPercentage ? 'RATE' : isCurrency;
-};
-
-export const dataTableChartNotesText =
-  'To optimize display performance, this chart represents data' +
-  ' points for the first record of each month plus the latest available record. The complete set' +
-  ' of data points for this date range can be found under the Table tab and are available through' +
-  ' the API endpoint for this data table.';
-
-const DataPreviewChart = ({ data, slug, currentTable, isVisible, legend, selectedPivot, dateField, dateRange, chartCitation }) => {
+const DataPreviewChart = ({ data, slug, currentTable, legend, selectedPivot, dateField }) => {
   const [chartFields, setChartFields] = useState([]);
   const [chartNotes, setChartNotes] = useState(null);
   const [hasUpdate, setHasUpdate] = useState(true);
   const [capitalized, setCapitalized] = useState('');
   const [axisHasBillions, setAxisHasBillions] = useState(false);
-
+  const [showLegend, setShowLegend] = useState(true);
   const viz = useRef();
 
   const buildLegendConfig = fields => {
@@ -116,14 +78,17 @@ const DataPreviewChart = ({ data, slug, currentTable, isVisible, legend, selecte
   const getVisibleChartFields = arr => arr.filter(f => f.active).map(ff => ff.field);
   const getActiveChartFields = arr => arr.filter(f => f.active);
   const activeChartFields = getActiveChartFields(chartFields);
-  useEffect(() => {
-    if (chartHooks && isVisible) {
-      chartHooks.onUpdateChartWidth(viz.current, activeChartFields, getVisibleChartFields(chartFields));
-    }
-  }, [legend, window.innerWidth]);
 
   useEffect(() => {
-    if (chartHooks && isVisible) {
+    if (chartHooks) {
+      console.log('here');
+      chartHooks.onUpdateChartWidth(viz.current, activeChartFields, getVisibleChartFields(chartFields));
+    }
+  }, [showLegend, window.innerWidth]);
+
+  useEffect(() => {
+    if (chartHooks) {
+      console.log('here 2');
       const nonActiveChartFields = getVisibleChartFields(chartFields);
       chartHooks.onUpdateChartWidth(
         viz.current,
@@ -132,7 +97,7 @@ const DataPreviewChart = ({ data, slug, currentTable, isVisible, legend, selecte
       );
       callbacks.onLabelChange(activeChartFields, chartFields, setChartFields);
     }
-  }, [isVisible]);
+  }, [chartHooks]);
 
   const determineIfAxisWillHaveBillions = data => {
     const valueArrays = data.map(v => Object.values(v));
@@ -145,7 +110,7 @@ const DataPreviewChart = ({ data, slug, currentTable, isVisible, legend, selecte
   useEffect(() => {
     let localChartFields;
 
-    if (isVisible && data && data.meta && !chartHooks) {
+    if (data && data.meta && !chartHooks) {
       localChartFields = setFieldsToChart(data.meta.dataTypes, selectedPivot, dateField);
 
       buildLegendConfig(localChartFields);
@@ -169,7 +134,7 @@ const DataPreviewChart = ({ data, slug, currentTable, isVisible, legend, selecte
         );
       }
     }
-  }, [isVisible, data]);
+  }, [data]);
 
   const handleLabelChange = update => {
     setHasUpdate(update.length > 0);
@@ -183,30 +148,34 @@ const DataPreviewChart = ({ data, slug, currentTable, isVisible, legend, selecte
   }, [selectedPivot]);
   ////${legend ? legendActive : ''}`}>
   return (
-    <div className={`${chartArea} ${chartFields.length <= 12 ? undefined : legendActive}`}>
+    <div className={`${chartArea} ${chartFields.length <= 12 || !showLegend ? undefined : legendActive}`}>
       <div className={chartPane}>
         <div className={chartLegendWrapper} style={chartFields.length <= 12 ? { flexDirection: 'column' } : { flexDirection: 'row' }}>
           <div className={vizClass}>
             {chartNotes}
             {selectedPivot && selectedPivot.pivotView?.roundingDenomination && (
               <div className={labelContainer}>
-                <div style={{ left: axisHasBillions && '-3rem' }} className={yAxisLabel}>
-                  {capitalized}
-                </div>
+                <div className={`${yAxisLabel} ${axisHasBillions ? billionsLabel : undefined}`}>{capitalized}</div>
               </div>
             )}
             <div id="viz" ref={viz} />
+            {chartFields.length > 13 && <ChartCitation slug={slug} currentTableName={currentTable.tableName} />}
           </div>
           {chartFields.length > 12 ? (
-            <div className={legendClass}>
-              <ChartLegendPanel
-                isVisible={isVisible}
-                fields={chartFields}
-                // If onHover is set to {callbacks.onHover}, then Jest can't tell onHover was fired.
-                onHover={(on, item) => callbacks.onHover(on, item, hasUpdate, chartFields)}
-                onLabelChange={handleLabelChange}
-              />
-            </div>
+            <>
+              <button className={legendToggle} onClick={() => setShowLegend(!showLegend)}>
+                X
+              </button>
+              <div className={legendClass}>
+                <ChartLegendPanel
+                  fields={chartFields}
+                  // If onHover is set to {callbacks.onHover}, then Jest can't tell onHover was fired.
+                  onHover={(on, item) => callbacks.onHover(on, item, hasUpdate, chartFields)}
+                  onLabelChange={handleLabelChange}
+                  legendVisibility={showLegend}
+                />
+              </div>
+            </>
           ) : (
             <ChartLegend
               fields={chartFields}
@@ -215,6 +184,7 @@ const DataPreviewChart = ({ data, slug, currentTable, isVisible, legend, selecte
               onLabelChange={handleLabelChange}
             />
           )}
+          {chartFields.length <= 12 && <ChartCitation slug={slug} currentTableName={currentTable.tableName} />}
         </div>
       </div>
     </div>
