+++ src/components/data-preview/data-preview-chart/data-preview-chart.module.scss
@@ -26,6 +26,10 @@ $paddingOffset: 87px;
   position: absolute;
   top: 11rem;
   left: -1rem;
+
+  .billionsLabel {
+    left: -3rem;
+  }
 }
 
 .datasetStats {
@@ -46,6 +50,7 @@ $paddingOffset: 87px;
 
 .legendActive {
   .legend {
+    width: $legendWidth;
     display: block;
   }
   .viz {
@@ -54,7 +59,8 @@ $paddingOffset: 87px;
 }
 
 .legend {
-  width: $legendWidth;
+  //display: none;
+  margin-top: 0.5rem;
   box-sizing: border-box;
   border: 1px solid $border-color;
   transition: width 250ms ease-in;
@@ -73,6 +79,12 @@ $paddingOffset: 87px;
   color: $info-icon;
   font-size: $font-size-20;
   margin-right: 1rem;
+}
+
+.legendToggle {
+  width: 24px;
+  height: 24px;
+  border-radius: 1rem;
 }
 
 @media screen and (max-width: $breakpoint-lg) {
