+++ src/components/data-preview/data-preview-filter-section/data-preview-table-filters/column-filter-options/date-column-filter/date-column-filter.tsx
@@ -8,10 +8,7 @@ const DateColumnFilter: FunctionComponent<IDateColumnFilter> = ({
   columnConfig,
   selectedTable,
   config,
-  setDateRange,
   allTablesSelected,
-  handleDateRangeChange,
-  setIsCustomDateRange,
   finalDatesNotFound,
   detailApi,
   detailViewState,
