+++ src/components/data-preview/data-preview-filter-section/data-preview-table-filters/data-preview-table-filters.tsx
@@ -315,10 +315,7 @@ const DataPreviewTableFilters: FunctionComponent<ITableFilters> = ({
       selectedColumn={selectedColumn}
       selectedTable={selectedTable}
       config={config}
-      setDateRange={setDateRange}
       allTablesSelected={allTablesSelected}
-      handleDateRangeChange={handleDateRangeChange}
-      setIsCustomDateRange={setIsCustomDateRange}
       finalDatesNotFound={finalDatesNotFound}
       detailApi={detailApi}
       detailViewState={detailViewState}
