+++ src/components/data-preview/data-preview-section-container/data-preview-section-container.module.scss
@@ -117,8 +117,7 @@ $table-container-margin: 8rem;
 }
 
 .noticeContainer {
-  padding: 0.5rem;
-  background-color: $body-background;
+  padding: 0 0 0.5rem 0;
 }
 
 .barExpander {
