+++ src/components/data-preview/data-preview.tsx
@@ -8,8 +8,6 @@ import {
   detailViewBack,
   detailViewButton,
   detailViewIcon,
-  placeholderButton,
-  placeholderText,
   selectedTableName,
   summaryTableHeader,
 } from './data-preview.module.scss';
@@ -108,6 +106,7 @@ export const DataPreview: FunctionComponent<IDataPreview> = ({
 
   const handleDateRangeChange = range => {
     if (range && isValidDateRange(range.from, range.to, config.techSpecs.earliestDate, config.techSpecs.latestDate)) {
+      console.log('here?');
       setDateRange(range);
     }
   };
@@ -137,9 +136,18 @@ export const DataPreview: FunctionComponent<IDataPreview> = ({
   }, [apis]);
 
   useEffect(() => {
+    console.log(selectedTable);
     if (selectedTable) {
       if (!selectedTable?.apiFilter?.disableDateRangeFilter) {
-        setDateRange(null);
+        // setDateRange(null);
+      } else if (selectedTable?.apiFilter?.disableDateRangeFilter) {
+        const defaultYear = new Date().getFullYear();
+        const defaultMonth = new Date().getMonth();
+        const startDate = new Date(defaultYear, defaultMonth, 1);
+        const endDate = new Date(defaultYear, defaultMonth + 1, 0);
+        // setSelectedMonth({ value: selectedTable?.apiFilter?.futureDated ? defaultMonth + 2 : defaultMonth + 1, label: monthFullNames[defaultMonth] });
+        // setSelectedYear({ value: defaultYear, label: defaultYear });
+        setDateRange({ from: startDate, to: endDate });
       }
       rewriteUrl(selectedTable, config.slug, location);
       setIsFiltered(true);
@@ -150,15 +158,6 @@ export const DataPreview: FunctionComponent<IDataPreview> = ({
       }
       setSelectedTableProp(selectedTable);
       // setYears(generateYearOptions(selectedTable?.earliestDate, selectedTable?.latestDate));
-      const defaultYear = new Date().getFullYear();
-      const defaultMonth = new Date().getMonth();
-      if (selectedTable?.apiFilter?.disableDateRangeFilter) {
-        const startDate = new Date(defaultYear, defaultMonth, 1);
-        const endDate = new Date(defaultYear, defaultMonth + 1, 0);
-        // setSelectedMonth({ value: selectedTable?.apiFilter?.futureDated ? defaultMonth + 2 : defaultMonth + 1, label: monthFullNames[defaultMonth] });
-        // setSelectedYear({ value: defaultYear, label: defaultYear });
-        setDateRange({ from: startDate, to: endDate });
-      }
     }
   }, [selectedTable]);
 
@@ -259,6 +258,10 @@ export const DataPreview: FunctionComponent<IDataPreview> = ({
 
   const checkDataDisplays = config.apis.every(api => (api?.dataDisplays?.length || 0) <= 1);
 
+  useEffect(() => {
+    console.log('dateRange', dateRange);
+  }, [dateRange]);
+
   return (
     <DatasetSectionContainer id="data-preview-table">
       <DataTableProvider config={config} detailViewState={detailViewState}>
@@ -339,18 +342,11 @@ export const DataPreview: FunctionComponent<IDataPreview> = ({
               currentDateButton={config.currentDateButton}
               datePreset={config.datePreset}
               customRangePreset={config.customRangePreset}
-              setIsFiltered={setIsFiltered}
               datasetDateRange={{
                 earliestDate: config.techSpecs.earliestDate,
                 latestDate: config.techSpecs.latestDate,
               }}
             >
-              {!selectedTable && (
-                <div data-testid="dateRangePlaceholder">
-                  <h3 className={placeholderText}>Date Range</h3>
-                  <div className={placeholderButton} />
-                </div>
-              )}
               {dateRange &&
                 (viewMode === 'table' ? (
                   <DataPreviewSectionContainer
@@ -392,13 +388,10 @@ export const DataPreview: FunctionComponent<IDataPreview> = ({
                   />
                 ) : (
                   <DataPreviewChart
-                    legend={false}
                     dateRange={dateRange}
                     data={apiData}
                     slug={config.slug}
                     currentTable={selectedTable}
-                    isVisible={true}
-                    chartCitation={false}
                     selectedPivot={selectedPivot}
                     dateField={dateFieldForChart}
                   />
